+++ CHANGELOG.rst
@@ -1,9 +1,9 @@
 Changelog
 =========
 
-x.y.z (MM/DD/YYYY)
+1.1.4 (9/29/2020)
 -----------------
-- Adjusted summarize to ignore phasing when creating consensus GT
+- Adjusted summarize to ignore phasing when creating consensus GT.
 - Adjusted Mutect handling to correctly marked phased normal (0|0)
   as non-somatic.
 - Removed extraneous fields from JQ_VS_LOW_CONFIDENCE filter metaheader.